fix(backend): convertir asyncpg.Record a dict en BaseRepository.fetch
Pydantic v2 con from_attributes=True no accede correctamente a los campos
de asyncpg.Record, reportando campos como 'missing' aunque estén presentes.
BaseRepository.fetch ahora retorna list[dict] para garantizar serialización
correcta en todos los endpoints de listado.

También corrige ProductOut.min_stock: int|None → Decimal|None ya que la
columna DB es NUMERIC y asyncpg retorna Decimal.

+++ backend/schemas/products.py
@@ -42,7 +42,7 @@ class ProductOut(BaseModel):
     price: Decimal | None
     cost: Decimal | None
     stock: Decimal
-    min_stock: int | None
+    min_stock: Decimal | None
     barcode: str | None
     sku: str | None
     is_variant: bool | None
